Lift constraints of SubstituteVisitor that tried to uphold validity of the dataflow tree
Lift some constraints of what can be substituted. These constraints where in place in an attempt to prevent erroneous modifications to the dataflow tree, resulting in invalid states.
This commit shifts the responsibility of keeping the dataflow graph in a valid state to the user allowing for a more flexible implementation.